requirements.txt
# Core Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0

# HuggingFace & Dataset Tools
datasets>=2.14.0
huggingface-hub>=0.19.0

# Vision-Language Models
pillow>=10.0.0
open-clip-torch>=2.23.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Evaluation & Metrics
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional but useful
ipython>=8.14.0
jupyter>=1.0.0
unzip
easydict
gdown